Reorganize lodestar/src/db directory
