--- packages/lodestar/src/db/api/index.ts
@@ -1,6 +1,0 @@
-/**
- * @module db/api
- */
-
-export {IBeaconDb} from "./beacon/interface";
-export {BeaconDb} from "./beacon/beacon";
