+++ packages/lodestar/src/db/beacon.ts
