+++ packages/lodestar/src/db/interface.ts
