+++ packages/lodestar/src/db/repositories/block/abstract.ts
