+++ packages/lodestar/src/db/repositories/blockArchive/abstract.ts
