+++ packages/lodestar/src/db/repositories/blockArchive/db-index.ts
