+++ packages/lodestar/src/db/repositories/blockArchive/index.ts
