+++ packages/lodestar/src/db/repositories/depositDataRoot.ts
