+++ packages/lodestar/src/db/repositories/index.ts
@@ -9,6 +9,7 @@ export * from "./attesterSlashing";
 export * from "./proposerSlashing";
 export * from "./voluntaryExit";
 export * from "./depositEvent";
+export * from "./pendingBlock";
 
 export * from "./depositDataRoot";
 export * from "./eth1Data";
