+++ packages/lodestar/src/db/repositories/stateArchive.ts
