+++ packages/lodestar/src/db/repositories/voluntaryExit.ts
