+++ packages/lodestar/src/db/single/preGenesisState.ts
