+++ packages/lodestar/src/db/single/preGenesisStateLastProcessedBlock.ts
